TASK: Remove unnecessary dependencies and latex support
Do not inherit from latex docker files. We do not support PDF in the
future. So we can keep the building and resulting docker image smaller.

Also we remove unnecessary dependencies like PHP or ncdu, which might be
useful during development of the image itself, but not for the user
afterwards.

Also documentation is adjusted to not mention unavailable support
anymore.